project2/requirements.txt
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
python-dotenv==1.2.2
requests==2.32.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
